CP-52793: remove sm-multipath, moved to host-installer
This script was only ever used by the host-installer when setting up
boot from SAN environments and this necessitated SM being included in
the install image for this one file. By moving this functionality into
the host-installer where it belongs we reduce the required
dependencies and size of the install image.

Signed-off-by: Mark Syms <mark.syms@cloud.com>