CA-395560 Improve logging and error checking on sg_readcap
When reading information using sg_readcap, it would be useful in the
event of failure to log what the return code was so that we know what
failure we're dealing with.

Also, per the manual pages for these commands, there are quite a few
more return codes which should be considered retryable, plus a non-zero
success return code which is theoretically possible.

Add a function to gather up the checks and logs for these things and
then wrap a call to that function in a short retry loop.

Signed-off-by: Tim Smith <tim.smith@cloud.com>