CP-49690 Create smlib to replace sm-core-libs
The sm-core-libs package mostly contained copies of libraries from SM,
but has the same LGPL License. It did install in the correct python
site-packages location, but it meant that any change to these libraries
had to be done in two different places which was both error- and
omission-prone.

This provides the same libraries that sm-core-libs provided, by moving
the SM originals from /opt/xensource/sm into the same place sm-core-libs
was using, so that sm-core-libs can be obsoleted.

It leaves behind in /opt/xensource/sm some wrapper libraries which
import the public APIs from the moved code, which should external code to
continue working until it can be updated. It also helps with the way
iscsilib was renamed to libiscsi by sm-core-libs.

Signed-off-by: Tim Smith <tim.smith@cloud.com>