CA-407507: Fix pvremove issue for multipathed devices
In the previous commit eb292457c5fd5f00f6fc82454a915068ab15aa6f,
we fixed an issue where the pvremove removed a device that
didn't represent the PV because of multipath.
We fixed that issue by finding the real PV before removing the VG and PVs.

Now another issue occured for a similar reason: after removing the VG,
LVM cache reassign the underlying device represents the PV, then we failed
`pvremove dev` using the original device node.
The issue can be reproduced repeatly on XenServer 9.
It might be caused by the update of lvm or multipath logic.

The solution is:
Before vgremove, save the uuid of the PVs attached to the VG.
After vgremove, use the uuid to get the underlying device that
represents the PV, then pvremove it.

For PV on multipathed devices, please refer to:
https://docs.redhat.com/en/documentation/red_hat_enterprise_linux/7/html/logical_volume_manager_administration/duplicate_pv_multipath#duplicate_pv_multipath

Signed-off-by: Stephen Cheng <stephen.cheng@cloud.com>