CA-407604: ensure that wwids are removed from /etc/multipath/wwids
Older versions of device-mapper-multipath would remove the wwid from
the /etc/multipath/wwids file when a `multipath -f` command is
issued. This is no longer the case and a `multipath -w`, which was
always available and is specifically for removing entries from this
file, is now required to do this.

Signed-off-by: Mark Syms <mark.syms@cloud.com>