CA-408452: remove vhd parent if it does not have one
A previous change stopped the cleanup code unilaterally removing the
`vhd-parent` configuration of the lef coalesced VHD which resulted in
failures where VHDs that should have declared a parent no longer did
until a subsequent SR scan restored the configuration value. This is
the alternate mirror case where the leaf coalesce collapses the entire
tree out and so the `vhd-parent` becomes redundant, again until a
subsequent SR scan removes it. To satisfy both use cases remove the
`vhd-parent` property if and only if the coalesced VHD has no parent.

Signed-off-by: Mark Syms <mark.syms@cloud.com>