Prevent fairlock from being stopped before xapi
When shutting down or rebooting, Xapi will call things in SM which want
to use the fairlock service, but newer systemd will refuse to start it
because it already has it stopping as part of the current transaction.

Ensure systemd does not stop fairlock until after Xapi, to prevent this
from being a problem.

Signed-off-by: Tim Smith <tim.smith@cloud.com>