CA-395221: use systemd target for gc enable
The xapi enabled state does not reflect whether the local xapi process
is able to respond to requests and only whether the host is enabled to
run new VMs. It is therefore not a useful check to ensure the GC will
be able to make requests of xapi. Switch instead to making the
service require the xapi-init-complete.target which will at least
ensure the local xapi has completed its initialisation which was the
original intent of this check.

Signed-off-by: Mark Syms <mark.syms@cloud.com>