CA-414438 Defend against unexpected VDI name_description contents
It seems to be possible for a description field on an LV snapshot to be
converted to a dictionary, which results in attempts to remove it from
the MGT volume failing after it is deleted, which in turn results in a
subsequent scan failing because an LV which should exist does not.

It is not immediately clear how this ended up as a dictionary as it has
been back and forth through XMLRPC a few times, but to solve the
immediate issue, defend against such things.

Signed-off-by: Tim Smith <tim.smith@cloud.com>