CA-418775: ensure getsize and sg_readcap agree on resize
When resizing the LUN underneath a PV for and SR ensure that the size
property in `/sys/block/<device>` and `sg_readcap` agree on the device
size before request `pvresize`. Intermittent failures have been
observered where `pvresize` would be called and not increase the size
of the PV. It appears that this has some relation to data in the
kernel device not seeing the update as being complete.

Signed-off-by: Mark Syms <mark.syms@citrix.com>