CA-421013: ensure cbt log removed on supporter after disable
When CBT is disabled the CBT log is deleted, on the coordinator, when
using shared LVM SRs it is necessary to inform the supporter of this
change so that it can remove all information relating to the CBT log
volume from the kernel device mapper table. If this is not done a
subsequent CBT enable for this VDI will fail if the device mapper
table still contains information about the previous, and now deleted,
volume.

Signed-off-by: Mark Syms <mark.syms@citrix.com>